Refactorment for Shared components. Moved them to a seperate folder
