+++ src/components/JobSearch/JobSearchForm.vue
@@ -35,7 +35,7 @@
 </template>
 
 <script>
-import ActionButton from "@/components/ActionButton.vue";
+import ActionButton from "@/components/Shared/ActionButton.vue";
 export default {
   name: "JobSearchForm",
   components: {
