+++ tests/unit/components/Shared/ActionButton.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/vue";
-import ActionButton from "@/components/ActionButton.vue";
+import ActionButton from "@/components/Shared/ActionButton.vue";
 
 describe("ActionButton", () => {
   it("renders text", () => {
